Drag window with the title bar (simplify version) the problems that i've countered today:
i want to move the window by grabing just only the title bar and not in any position of the window. but by clicking at one point of the "window" - #includedContent, the event only record at its click's element offset, and then move the window by the #includedContent offset defined by the click's element offset that is NOT from the #includedContent. Which resulted in a really poor glitch animation that maybe causing seizures. idk.

not only that, i can't handle the output of getElementById of an not-existing-yet element which is *seems like* very crucial in the project.

meh.